// src/app/shared/models/FirstAdminBoundary.model.ts
export interface IFirstAdminBoudary {
  OBJECTID: number;
  adm0_name: string;
  adm0_name_local: string;
  adm0_ISO3_2d: string;
  adm0_ISO3: string;
  validity: string;
  Shape__Area: number;
  Shape__Length: number;
  GlobalID: string;
  adm0_M49: string;
  pop_estim: null;
  pop_estim_source: null;
}