// src/app/shared/models/SecondAdminBoundary.model.ts
export interface ISecondAdminBoundary {
  OBJECTID: number;
  adm1_name: string;
  adm1_name_local: null;
  adm1_pcode: string;
  adm0_name: string;
  adm0_ISO3: string;
  validity: string;
  Shape__Area: number;
  Shape__Length: number;
  GlobalID: string;
  adm0_M49: string;
  pop_estim: null;
  pop_estim_source: null;
}