// src/app/shared/models/ThridAdminBoundary.model.ts
export interface IThirdAdminBoundary {
  OBJECTID: number;
  adm2_name: string;
  adm2_name_local: string;
  adm2_pcode: string;
  adm1_name: string;
  adm1_pcode: string;
  adm0_name: string;
  adm0_ISO3: string;
  validity: string;
  Shape__Area: number;
  Shape__Length: number;
  GlobalID: string;
  adm0_M49: string;
  pop_estim: null;
  pop_estim_source: null;
}